Refactor seed script to modularize asset seeding and remove redundant code
- Removed inline seeding logic for departments, users, locations, vendors, owners, categories, brands, models, assets, purchases, documents, assignments, maintenance tickets, disposals, software licenses, and custom statuses.
- Introduced a new function `seedAssets` to handle asset-related seeding.
- Simplified database connection setup and error handling for DATABASE_URL.
- Cleaned up unused imports and helper functions.
